Move EserialElement to model package
