+++ src/main/java/hu/elte/eserial/annotation/inclusionrule/AbstractInclusionRule.java
@@ -2,7 +2,7 @@ package hu.elte.eserial.annotation.inclusionrule;
 
 import hu.elte.eserial.annotation.EserialAnnotation;
 import hu.elte.eserial.annotation.enumeration.EserialAnnotationType;
-import hu.elte.eserial.recursion.model.EserialElement;
+import hu.elte.eserial.model.EserialElement;
 
 /**
  * An abstract class for evaluating {@link EserialAnnotation} {@link EserialAnnotationType#INCLUSION} constraints.
