+++ src/main/java/hu/elte/eserial/mapper/CompoundMapper.java
@@ -4,9 +4,9 @@ import hu.elte.eserial.annotation.UseToString;
 import hu.elte.eserial.annotation.processor.EserialAnnotationProcessor;
 import hu.elte.eserial.exception.EserialMapperMismatchException;
 import hu.elte.eserial.model.EserialContext;
+import hu.elte.eserial.model.EserialElement;
 import hu.elte.eserial.model.Getter;
 import hu.elte.eserial.recursion.RecursionChecker;
-import hu.elte.eserial.recursion.model.EserialElement;
 import hu.elte.eserial.util.AnnotationUtils;
 import hu.elte.eserial.util.MethodUtils;
 import hu.elte.eserial.util.TypeUtils;
