+++ src/test/java/hu/elte/eserial/annotation/processor/EserialAnnotationProcessorTest.java
@@ -4,7 +4,7 @@ import hu.elte.eserial.annotation.ExcludeThis;
 import hu.elte.eserial.annotation.IncludeElements;
 import hu.elte.eserial.annotation.IncludeMatching;
 import hu.elte.eserial.annotation.enumeration.IncludeMatcher;
-import hu.elte.eserial.recursion.model.EserialElement;
+import hu.elte.eserial.model.EserialElement;
 import org.junit.Before;
 import org.junit.Test;
