+++ src/test/java/hu/elte/eserial/mapper/MapperFactoryTest.java
@@ -1,6 +1,6 @@
 package hu.elte.eserial.mapper;
 
-import hu.elte.eserial.recursion.model.EserialElement;
+import hu.elte.eserial.model.EserialElement;
 import org.junit.Test;
 
 import java.util.Calendar;
