+++ src/test/java/hu/elte/eserial/recursion/RecursionCheckerTest.java
@@ -1,6 +1,6 @@
 package hu.elte.eserial.recursion;
 
-import hu.elte.eserial.recursion.model.EserialElement;
+import hu.elte.eserial.model.EserialElement;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -11,7 +11,9 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 public class RecursionCheckerTest {
