+++ src/test/java/hu/elte/eserial/recursion/model/EserialElementTest.java
@@ -1,5 +1,6 @@
 package hu.elte.eserial.recursion.model;
 
+import hu.elte.eserial.model.EserialElement;
 import org.junit.Before;
 import org.junit.Test;
